combined s3ImageRouter and sqlImageRouter files into one
